[inductor] Fix argmin/max with duplicate values
Fixes #99879

This adds `minimum_with_index` helper functions to compute the minimum
value and index simultaneously, with a preference for the smaller
index which is required to match eager in case of duplicates.

I also remove the mask-and-sum hack with a `tl.reduce` using
the previously mentioned helper. This additionally fixes the indices
being added together in the case of duplicates.

ghstack-source-id: 03dce80229d6b0180f307a3835f110f93de14d30
Pull Request resolved: https://github.com/pytorch/pytorch/pull/99920